Android Version / GUI
Made an android version which uses virtual joysticks to move and aim the
player, has buttons for attacking, blocking, sprinting.
Added some GUI images to display how much health the player has
remaining, and a menu which appears when your health reaches 0.
Allows the player to retry with a new dungeon or quit.